Move vars to a separate file from sprinkles
