+++ src/Playroom/CodeEditor/CodeEditor.css.ts
@@ -1,5 +1,6 @@
 import { style, globalStyle, keyframes, createVar } from '@vanilla-extract/css';
-import { vars, colorPaletteVars, sprinkles } from '../sprinkles.css';
+import { colorPaletteVars, sprinkles } from '../sprinkles.css';
+import { vars } from '../vars.css';
 import { toolbarItemSize } from '../ToolbarItem/ToolbarItem.css';
 
 const minimumLineNumberWidth = '50px';
