+++ src/Playroom/FramesPanel/FramesPanel.css.ts
@@ -1,6 +1,7 @@
 import { calc } from '@vanilla-extract/css-utils';
 import { globalStyle, style } from '@vanilla-extract/css';
-import { colorPaletteVars, sprinkles, vars } from '../sprinkles.css';
+import { colorPaletteVars, sprinkles } from '../sprinkles.css';
+import { vars } from '../vars.css';
 
 export const title = style({
   marginRight: 'auto',
