+++ src/Playroom/Snippets/Snippets.css.ts
@@ -1,6 +1,7 @@
 import { style } from '@vanilla-extract/css';
 import { toolbarItemSize } from '../ToolbarItem/ToolbarItem.css';
-import { sprinkles, vars, colorPaletteVars } from '../sprinkles.css';
+import { sprinkles, colorPaletteVars } from '../sprinkles.css';
+import { vars } from '../vars.css';
 
 export const root = sprinkles({
   position: 'relative',
